backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
PyPDF2==3.0.1
PyMuPDF==1.22.5
langchain==0.0.283
openai==1.3.5
chromadb==0.4.18
python-dotenv==1.0.0